pyproject.toml
[project]
name = "Android-SMS-API"
version = "0.1"
description = "Turn your Android phone into a programmable SMS server. A lightweight HTTP API wrapper around ADB for sending text messages over cellular network"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.4",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "jwt>=1.4.0",
    "uvicorn>=0.38.0",
    "python-multipart>=0.0.21",
    "bcrypt==3.2.0",
]
